+++ src/internals.ts
@@ -5,6 +5,8 @@
  */
 
 import { makeCtx } from "./ambient.ts";
+import { batch } from "./scheduling.ts";
+import { defer } from "./defer.ts";
 import { Job } from "./types.ts";
 
 export const
@@ -20,3 +22,8 @@ export const nullCtx = makeCtx();
 
 /** Jobs' owners (parents) - uses a map so child jobs can't directly access them */
 export const owners = new WeakMap<Job, Job>();
+
+/** Streams that need resuming  */
+export const pulls = /* @__PURE__ */ batch<{ doPull(): void; }>(pulls => {
+    for (const conn of pulls) { pulls.delete(conn); conn.doPull(); }
+}, defer);