+++ src/scheduling.ts
@@ -1,6 +1,4 @@
-import { type Cell } from "./cells.ts";
 import { defer } from "./defer.ts";
-import { DisposeFn } from "./types.ts";
 
 /** Scheduler state flags */
 const enum Is {
@@ -11,30 +9,85 @@ const enum Is {
 }
 
 /**
+ * A generic batch processing queue, implemented as a set.  (So items are
+ * processed at most once per batch run.)
+ *
+ * @template T The type of items that will be in the batch
+ *
+ * @category Scheduling
+ */
+export interface Batch<T> {
+    /** Is the batch processing function currently running? */
+    isRunning(): boolean;
+
+    /** Is the batch currently empty? */
+    isEmpty(): boolean;
+
+    /**
+     * Add an item to the batch.  Schedules the batch for processing if the
+     * batch was empty and is not already scheduled.  (Does nothing if the
+     * item is already in the batch.)
+     */
+    add(item: T): void;
+
+    /** Remove an item from the batch.  (Does not affect the schedule.) */
+    delete(item: T): void;
+
+    /** Is the item in the batch? */
+    has(item: any): boolean;
+
+    /**
+     * Process the batch now by calling the processing function, unless the
+     * batch is empty or already running.  If the processing function exits
+     * without fully emptying the batch (due to errors, time limits, etc.),
+     * another flush will be scheduled via the batch's scheduler (unless one is
+     * already scheduled).
+     */
+    readonly flush: () => void;
+}
+
+
+/**
+ * Create a batch processing queue from the given processing loop and scheduling
+ * function.
+ *
+ * @template T The type of items that will be in the batch
+ *
+ * @param process A function taking a set of items.  It should remove items from
+ * the set, usually *before* processing them.  (So that the batch won't
+ * perpetually block on that item in case of a persistent error.)  If the
+ * processing function doesn't remove all items from the set, another processing
+ * pass will be done later.  (This allows you to rate-limit processing, so as
+ * not to hog the current thread.)
+ *
+ * @param sched A single-argument scheduling function (like
+ * requestAnimationFrame, setImmediate, or queueMicrotask).  The batch will call
+ * it from time to time with a single callback.  The scheduling function should
+ * then arrange for that callback to be invoked *once* at some future point,
+ * when it is the desired time for pending items to be processed.  If no
+ * function is given (or it's undefined/null), {@link defer} is used.
+ *
+ * @returns a {@link Batch} that items can be added to for later processing.
+ *
+ * @category Scheduling
+ */
+export function batch<T>(process: (items: Set<T>) => void, sched?: (cb: () => unknown) => unknown): Batch<T> {
+    return new _Batch(process, sched);
+}
+
+/**
  * A generic queue for things that need to be run at-most-once per item
  * per timeframe.
  *
- * @category Jobs and Scheduling
+ * @category Scheduling
  */
-export class RunQueue<K> {
+class _Batch<T> implements Batch<T> {
     protected _flags: Is = Is.Unset;
-    protected readonly q = new Set<K>();
+    protected readonly q = new Set<T>();
 
     constructor(
-        /**
-         * A single-argument scheduling function (like requestAnimationFrame,
-         * setImmediate, or queueMicrotask).  The scheduler will call it from
-         * time to time with a single callback.  The scheduling function should
-         * then arrange for that callback to be invoked *once* at some future
-         * point, when it is the desired time for all pending items on this
-         * queue to run.
-         */
-        protected readonly sched: (cb: () => unknown) => unknown,
-        /**
-         * A callback to loop over the queue, removing items and performing any
-         * necessary operations
-         */
-        protected readonly reap: (queue: Set<K>) => void
+        protected readonly reap: (queue: Set<T>) => void,
+        protected readonly sched: ((cb: () => unknown) => unknown) | undefined | null,
     ) {}
 
     /** Is this queue currently running? */
@@ -43,7 +96,7 @@ export class RunQueue<K> {
     /** Is this queue currently empty? */
     isEmpty() { return !this.q.size; }
 
-    add(subject: K) {
+    add(subject: T) {
         this.q.size || this._flags & (Is.Running|Is.Scheduled) || this._sched();
         this.q.add(subject);
     }
@@ -58,7 +111,7 @@ export class RunQueue<K> {
 
     protected _sched() {
         this._flags |= Is.Scheduled;
-        this.sched(this._run);
+        (this.sched || defer)(this._run);
     }
 
     protected _run = () => {
@@ -83,43 +136,3 @@ export class RunQueue<K> {
         }
     }
 }
-
-export var currentRule: Cell;
-var currentQueue: RuleQueue;
-
-const ruleQueues = new WeakMap<Function, RuleQueue>();
-export const ruleStops = new WeakMap<Cell, DisposeFn>();
-
-export class RuleQueue extends RunQueue<Cell> {
-    constructor(sched: (cb: () => unknown) => unknown) {
-        super(sched, function(this: RuleQueue, q) {
-            // another queue is running? reschedule for later
-            if (currentQueue) return;
-            currentQueue = this;
-            try {
-                // run rules marked dirty by value changes
-                for (currentRule of q) {
-                    currentRule.catchUp();
-                    q.delete(currentRule);
-                }
-            } finally {
-                currentQueue = currentRule = undefined;
-            }
-        });
-    }
-}
-
-export function ruleQueue(scheduleFn: (cb: () => unknown) => unknown = defer) {
-    ruleQueues.has(scheduleFn) || ruleQueues.set(scheduleFn, new RuleQueue(scheduleFn));
-    return ruleQueues.get(scheduleFn);
-}
-
-export const defaultQ = /* @__PURE__ */ ruleQueue(defer);
-
-/** @internal */
-export const pulls = /* @__PURE__ */ new RunQueue<{doPull(): void}>(defer, pulls => {
-    for (const conn of pulls) { pulls.delete(conn); conn.doPull(); }
-})
-
-/** @internal For testing only */
-export const runPulls  = /* @__PURE__ */ (() => pulls.flush)();